requirements.txt
flask
flask_cors
numpy
Pillow
mxnet
opencv-python
sklearn
scikit-learn
scikit-image